% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalreg.R
\name{generalreg}
\alias{generalreg}
\title{generalreg}
\usage{
generalreg(
  data,
  mu_formula,
  var_formula = NULL,
  dist = "normal",
  alpha = NULL,
  beta = NULL,
  start = NULL
)
}
\arguments{
\item{data}{A data frame in which to evaluate the variables in \code{formula} and \code{formula_var}.
Can also be a list or an environment, but not a matrix}

\item{mu_formula}{a model formula including variables and parameters}

\item{var_formula}{a  model formula for the diagonal of covariance matrix}

\item{dist}{Inform the distribution of your data, can be: "normal", "logistic", "t" or "power exp"}
}
\value{
generalreg returns an object of class `generalreg`
}
\description{
generalreg
}
\examples{
library(generalreg)
set.seed(3)
X <- data.frame(x1 = rnorm(1000), x2 = rnorm(1000), x3 = rnorm(1000))
e = rt(1000, df = 3)
e = rlogis(10000, scale = 2, location = 3)
e = rnorm(1000, sd = 5)
y <- 2 + 3 * X$x1 + 7 * X$x2 + e
data <- data.frame(y, X)
generalreg(data, mu_formula = y ~ beta0 + beta1 * x1 + beta2 * x2, dist='normal')


y <- (1 / 2 * X$x1) + e
data <- data.frame(y, X)
generalreg(data, mu_formula = y ~ beta0 / beta1 * x1  , dist='normal')

generalreg(data = mtcars, mu_formula = mpg ~ alfa + 1/ (beta * disp))


}
